+++ src/pharmpy/cli.py
@@ -644,7 +644,10 @@ def results_frem(args):
     res.add_plots()
     res.to_json(path=args.psn_dir / 'results.json')
     res.to_csv(path=args.psn_dir / 'results.csv')
-    res.create_report(args.psn_dir)
+
+    from pharmpy.modeling.reporting import create_report
+
+    create_report(res, args.psn_dir)
 
 
 def results_linearize(args):
@@ -731,7 +734,9 @@ def results_report(args):
     from pharmpy.results import read_results
 
     res = read_results(results_path)
-    res.create_report(args.psn_dir)
+    from pharmpy.modeling.reporting import create_report
+
+    create_report(res, args.psn_dir)
 
 
 def results_summary(args):