+++ tests/tools/test_frem.py
@@ -538,7 +538,7 @@ def test_modeling_create_results(testdata):
 def test_create_report(testdata, tmp_path):
     res = modeling.read_results(testdata / 'frem' / 'results.json')
     shutil.copy(testdata / 'frem' / 'results.json', tmp_path)
-    res.create_report(tmp_path)
+    modeling.create_report(res, tmp_path)
     html = tmp_path / 'results.html'
     assert html.is_file()
     assert html.stat().st_size > 500000